Refactor * Changed simple_model architecture * Change episode to 10 games instead of 1 * Added some perf counter * Eval perf     * Model evaluation and categorical distribution takes majority       of time     * Added multiprocess and gpu support. Does not fully work

+++ connect4/board.py
@@ -11,7 +11,7 @@ class Board:
         self.reset()
 
     def reset(self):
-        self.cells = np.full((self.height, self.width), Rules.ColorNone, dtype=np.float)
+        self.cells = np.full((self.height, self.width), Rules.ColorNone, dtype=np.float32)
         self.heights = np.zeros(self.width, dtype=np.int)
         self.occupiedCells = 0
